+++ src/main/java/com/example/yourmode/domain/member/entity/LoginType.java
@@ -1,4 +1,4 @@
-package com.example.yourmode.domain.member.domain;
+package com.example.yourmode.domain.member.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
