+++ src/main/java/com/example/yourmode/domain/member/repository/MemberRepository.java
@@ -1,10 +1,8 @@
 package com.example.yourmode.domain.member.repository;
 
-import com.example.yourmode.domain.member.domain.Member;
-import com.example.yourmode.domain.member.domain.LoginType;
+import com.example.yourmode.domain.member.entity.Member;
+import com.example.yourmode.domain.member.entity.LoginType;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
