+++ src/main/java/com/example/yourmode/domain/member/service/MemberRefreshTokenService.java
@@ -1,6 +1,6 @@
 package com.example.yourmode.domain.member.service;
 
-import com.example.yourmode.domain.member.domain.Member;
+import com.example.yourmode.domain.member.entity.Member;
 import com.example.yourmode.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
